test: move ECDSA activation tests into activation_test
